Character Control Origins
Limited movement only possible on your "Turn"
(turn regenerates over time in this version, but later will be
incremented by a game controller, or the enemies ending their turns)

Press 'A' to activate movement mode. (The cursor will appear)

While in movement mode, press "A" again to move to the cursor position,
or press "Space" to exit movement mode.