Added New Enemy: Moblin With Animations
I added a Moblin as a new enemy. I created animations for when it is
moving and when it is at idle. There is a corresponding Moblin script
where the enemy logic is located. To promote organization, I added some
folders to the existing Prefabs and Animation folders. I configured it
to move in real time while testing.