Added 10 unique maps, level selector function, and a pit tile
+Added Pit Tile
+Added 10 Unique maps (BoardManager.cs)
+Added Level Selector in BoardManager.cs
+Added variable 'level' in GameManager.cs
(Note: this will be used for telling the board what level to load)

As of right now,  there are 10 levels each with a ladder to proceed to
the next level. At the moment, the ladder will not transition the player
to the next level (I plan on putting that in Thursday or Friday). So if
you want to check each level design, you have to manually change the
'level' in GameManager.cs