Attempted Level Transitions, Had issues with Player not moving
+Added new functions to GameManager.cs
OnLevelWasLoaded should be called each time the level loads and it
increases the level counter to move the board to the next map then moves
on to Init. This one just tells the board to load the map according to
the level.

+Added 2 functions in GridAura.cs (I think this is where I should put
them, I could be wrong)  I put in OnTriggerEnter2D which checks if the
tile it collided with has the Exit tag and if it does it calls Restart.
Restart reloads the scene and it should then increment the level using
the new functions above.

At the moment, I could not test if these work due to an issue with the
player being unable to move