Basic Items, Leveling System, & Stat Logic Complete
All of my parts for sprint 2 should be completed now. The items can now
be picked up and are added to the user’s inventory. Right now I set it
up where if an enemy comes into contact with an enemy they defeat it
and get exp. The stat system is also set up for the player and the
moblin.