Attack and Defense Stat Now Impact Damage Amount
The attack and defense stats now influence the amount of damage that an
enemy or player can deal. The difference between the stats must be at
least 2 for additional damage to occur. For example; If the player’s
attack is 3 and the enemy’s defense is 1, then the player will cause
the enemy’s health to drop by 2.