Completed Difficulty (Normal | Hard) Logic & Sword Damage Logic and Animations
I have completed both of my tasks (plus others that I needed to have
before working on my tasks). The player can now receive damage from the
Moblin and will be destroyed when their health <= 0. The same logic
applies to the enemy moblin. Furthermore, I added attack states to the
moblin. These are activated when the moblin attacks the enemy.